+++ src/Row.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 
 import Cell from './Cell';
 import { RowRendererProps } from './common/types';
+import { preventDefault, wrapEvent } from './utils';
 
 export default function Row<R>({
   cellRenderer: CellRenderer = Cell,
@@ -10,37 +11,28 @@ export default function Row<R>({
   enableCellRangeSelection,
   eventBus,
   height,
-  idx,
+  rowIdx,
   isRowSelected,
   isSummaryRow,
   lastFrozenColumnIndex,
-  onDragEnter,
-  onDragOver,
-  onDrop,
   row,
   scrollLeft,
   viewportColumns,
   width,
+  onDragEnter,
+  onDragOver,
+  onDrop,
   ...props
 }: RowRendererProps<R>) {
   function handleDragEnter(event: React.DragEvent<HTMLDivElement>) {
     // Prevent default to allow drop
     event.preventDefault();
-    eventBus.dispatch('DRAG_ENTER', idx);
-    onDragEnter?.(event);
+    eventBus.dispatch('DRAG_ENTER', rowIdx);
   }
 
   function handleDragOver(event: React.DragEvent<HTMLDivElement>) {
     event.preventDefault();
     event.dataTransfer.dropEffect = 'copy';
-    onDragOver?.(event);
-  }
-
-  function handleDrop(event: React.DragEvent<HTMLDivElement>) {
-    // The default in Firefox is to treat data in dataTransfer as a URL and perform navigation on it, even if the data type used is 'text'
-    // To bypass this, we need to capture and prevent the drop event.
-    event.preventDefault();
-    onDrop?.(event);
   }
 
   function getCells() {
@@ -49,7 +41,7 @@ export default function Row<R>({
         <CellRenderer
           key={column.key as string} // FIXME: fix key type
           idx={column.idx}
-          rowIdx={idx}
+          rowIdx={rowIdx}
           column={column}
           lastFrozenColumnIndex={lastFrozenColumnIndex}
           row={row}
@@ -65,18 +57,21 @@ export default function Row<R>({
 
   className = classNames(
     'rdg-row',
-    `rdg-row-${idx % 2 === 0 ? 'even' : 'odd'}`,
+    `rdg-row-${rowIdx % 2 === 0 ? 'even' : 'odd'}`,
     { 'rdg-row-selected': isRowSelected },
     className
   );
 
+  // Regarding onDrop: the default in Firefox is to treat data in dataTransfer as a URL,
+  // and perform navigation on it, even if the data type used is 'text'.
+  // To bypass this, we need to capture and prevent the drop event.
   return (
     <div
       className={className}
       style={{ width, height }}
-      onDragEnter={isSummaryRow ? onDragEnter : handleDragEnter}
-      onDragOver={isSummaryRow ? onDragOver : handleDragOver}
-      onDrop={isSummaryRow ? onDrop : handleDrop}
+      onDragEnter={isSummaryRow ? onDragEnter : wrapEvent(handleDragEnter, onDragEnter)}
+      onDragOver={isSummaryRow ? onDragOver : wrapEvent(handleDragOver, onDragOver)}
+      onDrop={isSummaryRow ? onDrop : wrapEvent(preventDefault, onDrop)}
       {...props}
     >
       {getCells()}