+++ stories/index.tsx
@@ -12,6 +12,8 @@ import NoRows from './demos/NoRows';
 import TreeView from './demos/TreeView';
 import ContextMenu from './demos/ContextMenu';
 import ScrollToRow from './demos/ScrollToRow';
+import CellNavigation from './demos/CellNavigation';
+import HeaderFilters from './demos/HeaderFilters';
 
 storiesOf('Demos', module)
   .add('Common Features', () => <CommonFeatures />)
@@ -21,4 +23,6 @@ storiesOf('Demos', module)
   .add('Cell Actions', () => <CellActions />)
   .add('Tree View', () => <TreeView />)
   .add('Context Menu', () => <ContextMenu />)
-  .add('Scroll To Row', () => <ScrollToRow />);
+  .add('Scroll To Row', () => <ScrollToRow />)
+  .add('Cell Navigation', () => <CellNavigation />)
+  .add('Header Filters', () => <HeaderFilters />);